% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dbc_sample_pasta}
\alias{dbc_sample_pasta}
\title{Internal test data for dtsus package}
\description{
This dataset is used internally by the test suite.
It is loaded from \code{data/dbc_teste.rda} and contains
sample DBC file information for testing.
}
\keyword{internal}
