% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtsus_pop_ans.R
\name{dtsus_pop_ans}
\alias{dtsus_pop_ans}
\title{Populacao beneficiaria de planos de saude da ANS}
\usage{
dtsus_pop_ans(ano_mes, uf, quiet = FALSE)
}
\arguments{
\item{ano_mes}{Competencia no formato \code{"AAAAMM"}.
Exemplo: \code{"202403"} para marco de 2024.}

\item{uf}{Sigla da Unidade Federativa. Exemplo: \code{"MG"}, \code{"SP"}, \code{"BA"}.}

\item{quiet}{Logico. Se \code{TRUE}, oculta as mensagens e a barra de progresso
do download. O padrao e \code{FALSE}.}
}
\value{
Um \code{data.frame} contendo os dados de beneficiarios da ANS
para a UF e competencia informadas.
}
\description{
Baixa, extrai e le a base de informacaes consolidadas de
beneficiarios de planos de saude disponibilizada pela Agencia
Nacional de Saude Suplementar (ANS).
}
\examples{
\dontrun{
ans <- dtsus_pop_ans("202403", "MG", quiet = FALSE)
}

}
