% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtsus_download.R
\name{dtsus_download}
\alias{dtsus_download}
\title{Fetch and read microdata files from DATASUS}
\usage{
dtsus_download(
  fonte = NA,
  tipo = NA,
  uf = NA,
  Data_inicio = 202308,
  Data_fim = 202512,
  origem = "data.sus"
)
}
\arguments{
\item{fonte}{A string. The abbreviation of the health information system to be accessed.}

\item{tipo}{A string. The abbreviation of the file to be accessed.}

\item{uf}{A string or a vector of strings. By default ...}

\item{Data_inicio}{Start year and month in the format yyyymm.}

\item{Data_fim}{End year and month in the format yyyymm.}

\item{origem}{I don't know..}
}
\value{
fazer
}
\description{
Fetch and read microdata files from DATASUS
}
\examples{
fazer


}
